Refactor and Create New Hooks
